+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 1.30.0 - 16/03/2026
+* Ajustes nas notas dos pedidos.
+* Ajuste no reembolso do cartão de crédito e débito.
+* Novo sistema de captura parcial.
+* Ajuste na validação do cartão que inicia com 0.
+
 # 1.29.0 - 04/03/2026
 * Novo layout para o shortcode.
 * Ajustes de CSS na versão padrão e moderna.